requirements.txt
fastapi==0.115.8
pydantic==2.10.6
pydantic-settings==2.8.0
email_validator==2.2.0
uvicorn==0.34.0
SQLAlchemy==2.0.38
redis==5.2.1
asyncpg==0.30.0
alembic==1.14.1
logger==1.4
passlib==1.7.4
python-jose==3.4.0
requests==2.32.3
